+++ src/main/java/WishPool/Be/file/exception/GcsUploadException.java
@@ -1,4 +1,4 @@
-package WishPool.Be.gcs.exception;
+package WishPool.Be.file.exception;
 
 // GCS 업로드 중 발생하는 예외임을 명확히 하기 위한 클래스
 public class GcsUploadException extends RuntimeException {
