+++ src/main/java/WishPool/Be/jwt/GeneratedToken.java
@@ -1,4 +1,4 @@
-package WishPool.Be.util.jwt;
+package WishPool.Be.jwt;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
